refactor!: restructure project
