+++ .eslintignore
@@ -37,4 +37,4 @@ yarn-error.log*
 *.tsbuildinfo
 
 # generated files
-src/types/i18n.ts
+/src/modules/i18n/types/i18n.ts
