+++ src/modules/i18n/index.ts
@@ -1,0 +1,4 @@
+import useTranslation from "./hooks/use-translation";
+import type I18nKey from "./types/i18n";
+
+export { I18nKey, useTranslation };
