+++ src/modules/theme/index.ts
@@ -1,0 +1,4 @@
+import useTheme from "./hooks/use-theme";
+import type Theme from "./types/theme";
+
+export { Theme, useTheme };
